Refactor backend folder structure, improving organization
