+++ backEnd/index.js
@@ -4,17 +4,16 @@ const bodyParser = require('body-parser');
 const session = require('express-session');
 const cors = require('cors');
 const FileStore = require('session-file-store')(session);
-const homeRoute = require('./private/src/routing/home.js');
-const teamsRoutes = require('./private/src/routing/teams.js');
-const teamRoutes = require('./private/src/routing/team.js');
-const playerRoutes = require('./private/src/routing/player.js');
-const resetRoutes = require('./private/src/routing/reset.js');
-const userSessionRoutes = require('./private/src/routing/user.js');
-const errorRoutes = require('./private/src/routing/error.js');
-const teamCrestRoutes = require('./private/src/routing/crest.js');
-const { paths, CLIENT_BASE_URL } = require('./private/src/routing/paths.js');
+const teamsRoutes = require('./src/components/routing/teams.js');
+const teamRoutes = require('./src/components/routing/team.js');
+const playerRoutes = require('./src/components/routing/player.js');
+const resetRoutes = require('./src/components/routing/reset.js');
+const userSessionRoutes = require('./src/components/routing/user.js');
+const errorRoutes = require('./src/components/routing/error.js');
+const teamCrestRoutes = require('./src/components/routing/crest.js');
+const { paths, CLIENT_BASE_URL } = require('./src/components/routing/paths.js');
 
-const { ensureLoggedIn } = require('./private/src/auth.js');
+const { ensureLoggedIn } = require('./src/components/auth.js');
 
 const PORT = 3000;
 const app = express();
@@ -22,8 +21,8 @@ const corsOptions = {
   origin: CLIENT_BASE_URL,
   credentials: true,
 };
-app.use(express.static(path.join(__dirname, 'public')));
-const fileStoreOptions = { path: path.join(__dirname, 'private', 'sessions'), ttl: 60 * 60 * 24 * 7, logFn: () => {} };
+//app.use(express.static(path.join(__dirname, 'src', 'userData')));
+const fileStoreOptions = { path: path.join(__dirname, 'src', 'sessions'), ttl: 60 * 60 * 24 * 7, logFn: () => {} };
 app.use(session({
   store: new FileStore(fileStoreOptions),
   secret: 'keyboard-cat',
@@ -42,7 +41,6 @@ app.use((req, res, next) => {
 });
 app.use(bodyParser.json());
 
-app.use(paths.home, homeRoute);
 app.use(paths.teams, teamsRoutes);
 app.use(paths.player, playerRoutes);
 app.use(paths.team, teamRoutes);
