+++ backEnd/src/components/user.js
@@ -4,13 +4,19 @@ const { getUserRootPath, getUserTeamsFolderPath, getUserCustomCrestFolderPath } 
 
 function createUser(username) {
   try {
+    console.log(`Creating new user ${username}`);
+    console.log(getUserRootPath(username));
+    console.log(getUserTeamsFolderPath(username));
+    console.log(getUserCustomCrestFolderPath(username));
+
+
     createFolder(getUserRootPath(username));
     createFolder(getUserTeamsFolderPath(username));
     createFolder(getUserCustomCrestFolderPath(username));
     const defaultUsername = 'default';
     copyTeamList(defaultUsername, username);
-  } catch {
-    throw new Error('Failed to create new user');
+  } catch (e) {
+    throw new Error('Failed to create new user: ' + e);
   }
 }
 function deleteUser(username) {
